:bug: change price type from float64 to uint

+++ entities/item.go
@@ -13,7 +13,7 @@ type (
 		Name        string    `gorm:"type:varchar(64);unique;not null;"`
 		Description string    `gorm:"type:varchar(128);not null;"`
 		Picture     string    `gorm:"type:varchar(256);not null;"`
-		Price       float64   `gorm:"not null;"`
+		Price       uint      `gorm:"not null;"`
 		IsArchive   bool      `gorm:"not null;default:false;"`
 		CreatedAt   time.Time `gorm:"not null;autoCreateTime;"`
 		UpdatedAt   time.Time `gorm:"not null;autoUpdateTime;"`
