+++ pkg/itemShop/model/itemShop.go
@@ -2,11 +2,11 @@ package model
 
 type (
 	Item struct {
-		ID          uint64  `json:"id"`
-		Name        string  `json:"name"`
-		Description string  `json:"description"`
-		Picture     string  `json:"picture"`
-		Price       float64 `json:"price"`
+		ID          uint64 `json:"id"`
+		Name        string `json:"name"`
+		Description string `json:"description"`
+		Picture     string `json:"picture"`
+		Price       uint   `json:"price"`
 	}
 
 	ItemFilter struct {
